CPPCheckBearTest.py: Convert test to new style
Converted to ease the migration of the bear to a global bear.
