+++ tests/c_languages/cppcheck_test_files/bad_file.cpp
@@ -1,0 +1,8 @@
+#define f(c) { \
+    char s[10]; \
+    s[c] = 42; \
+}
+int main() {
+    f(100);
+    return 0;
+}
