+++ tests/c_languages/cppcheck_test_files/good_file.cpp
@@ -1,0 +1,5 @@
+using namespace std;
+int main() {
+    cout << "Hello, world!" << endl;
+    return 0;
+}
