+++ tests/c_languages/cppcheck_test_files/warn_file.cpp
@@ -1,0 +1,6 @@
+void f1(struct fred_t *p)
+{
+    int x;
+    if (p)
+        do_something(x);
+}
